Update plugins to export a factory function
