+++ packages/anvil-plugin-code-splitting/src/index.ts
@@ -1,37 +1,4 @@
-import babelPreset from './babel'
-import { HandlerArgs } from '@financial-times/anvil'
+import { plugin } from './plugin'
 
-export default ({ on }) => {
-  on('babelConfig', amendBabelConfig)
-  on('webpackConfig', getWebpackConfigToMerge)
-}
-
-function amendBabelConfig({ cli, resource: babelConfig }: HandlerArgs) {
-  babelConfig.presets.push(babelPreset(cli))
-}
-
-function getWebpackConfigToMerge() {
-  return {
-    optimization: {
-      runtimeChunk: 'single',
-      splitChunks: {
-        chunks: 'all',
-        maxInitialRequests: Infinity,
-        minSize: 0,
-        cacheGroups: {
-          vendor: {
-            test: /[\\/]node_modules[\\/]/,
-            name(module) {
-              // get the name. E.g. node_modules/packageName/not/this/part.js
-              // or node_modules/packageName
-              const packageName = module.context.match(/[\\/]node_modules[\\/](.*?)([\\/]|$)/)[1]
-
-              // npm package names are URL-safe, but some servers don't like @ symbols
-              return `npm.${packageName.replace('@', '')}`
-            }
-          }
-        }
-      }
-    }
-  }
-}
+export { plugin }
+export default plugin
