+++ packages/anvil-plugin-esnext/src/plugin.ts
@@ -1,0 +1,12 @@
+import babelPreset from './babel'
+import { HandlerArgs } from '@financial-times/anvil'
+
+export function plugin() {
+  return ({ on }) => {
+    on('babelConfig', amendBabelConfig)
+  }
+}
+
+function amendBabelConfig({ cli, resource: babelConfig }: HandlerArgs) {
+  babelConfig.presets.push(babelPreset(cli))
+}
