+++ packages/anvil-plugin-ft-css/src/index.ts
@@ -1,4 +1,5 @@
 import { plugin } from './plugin'
 
 export { hooks } from './hooks'
+export { plugin }
 export default plugin
